Eliminate redundant server-side copy in get_blob tests.
The get_blob() function in target_code.{h,cc} carefully caches strings of various sizes, but then returns a copy of the cached string rather than a reference. I assume the intent was *not* to make a copy here, since otherwise the caching does not make sense.

(Note: Historically, std::strings were reference-counted, in which case no copy would happen here. However, C++11 forbids refcounting strings because the atomic operations turned out to cost more than they saved. As of GCC 5, std::string is not refcounted.)

With this change, the "direct" call now takes negligible time. It's unclear whether this should be considered "correct", since "direct" is already a somewhat degenerate case.

Aside from "direct", it looks like this redundant copy was hitting all the benchmark fixtures equally, so it was not "unfair" per se, but it was diluting the test results somewhat.