Remove redundant copy from capnp get_blob client.
The capnp test fixture was copying the returned blob from getBlob() into an std::string, rather than use the capnp::Data::Reader directly. For gRPC, Thrift, and rpclib, on the other hand, the client was using the std::string reference provided by the library without making any copies.

It's unfair to force a random extra copy on Cap'n Proto, so I've changed the code to match more closely what the other fixtures do.

(This also happens to delete a bug: the std::string was being constructed from a char* pointer without passing the size, meaning it was assumed to be NUL-terminated. However, the actual data could contain NUL characters. So the copy was averaging 256 bytes anyway, although it could have segfaulted if the data happened to contain no NUL bytes.)

One could argue that it is "fair" to force this extra copy on Cap'n Proto in the case that the client app is designed to operate on std::strings and cannot accept a byte pointer/size pair. However:

1. In real life, apps very often can accept a pointer/size pair, especially if they expect to deal in large blobs.
2. Apps which do demand an std::string often need to make a copy even when using an RPC library that provides std::string, due to ownership issues.
3. Cap'n Proto goes to great lengths to avoid extra copies of data, and one way it does that is by providing a pointer into the original message buffer rather than copy into an std::string; it should get credit for providing a performance feature that other RPC libraries lack.